mod - consistent, typed functions to talked to graphql.
