+++ src/index.ts
@@ -6,9 +6,15 @@ import axios from 'axios';
 
 import * as _ from 'lodash';
 
-import { applyMiddleware, applyRoutes, Route } from "./utils";
+import { applyMiddleware, applyRoutes, contentTypeHeaders, graphqlEndpoint, Route } from "./utils";
 import * as utils from "./utils";
 import * as middleware from "./middleware";
+
+import { askBestBlock } from "./services/bestblock";
+import { askUtxoForAddresses } from "./services/utxoForAddress";
+import { askBlockNumByHash, askBlockNumByTxHash, askTransactionHistory } from "./services/transactionHistory";
+import { askFilterUsedAddresses } from "./services/filterUsedAddress";
+import { askUtxoSumForAddresses } from "./services/utxoSumForAddress";
 
 
 const router = express();
@@ -20,304 +26,29 @@ const middlewares = [ middleware.handleCors
 
 applyMiddleware(middlewares, router);
 
-const graphqlEndpoint = 'http://localhost:3100/graphq';
 const port = 8082;
 const addressesRequestLimit = 50;
 const apiResponseLimit = 50; 
 
-
-const contentTypeHeaders = { headers: {'Content-Type': 'application/json'}};
-const askBestBlock = () => {
-    const query = `
-                {
-                  cardano {
-                    blockHeight,
-                    currentEpoch {
-                      number
-                      blocks(limit:1, order_by: { createdAt:desc}) {
-                        hash
-                        number
-                      }
-                    },
-                    slotDuration,
-                  },
-                }
-            `;
-    return axios.post(graphqlEndpoint, JSON.stringify({'query':query}), contentTypeHeaders);
-};
-
 const bestBlock = async (req: Request, res: Response) => {
-  const { cardano } = (await askBestBlock()).data.data;
-  res.send({
-    epoch: cardano.currentEpoch.number,
-    slot: cardano.slotDuration,
-    hash: cardano.currentEpoch.blocks[0].hash,
-    height: cardano.blockHeight,
-  });
-};
+  const result = await askBestBlock();
+  switch(result.kind) {
+    case "ok":
+      const cardano = result.value;
+      res.send({
+        epoch: cardano.currentEpoch.number,
+        slot: cardano.slotDuration,
+        hash: cardano.currentEpoch.blocks[0].hash,
+        height: cardano.blockHeight,
+      });
 
-const askFilterUsedAddresses = (addresses: string[]) => {
-    const query = `
-                    query UsedAddresses($addresses: [String!]) {
-                      transactions(where: {
-                        _or: [
-                          {
-                            inputs: {
-                              address: {
-                                _in: $addresses
-                            }
-                          }},
-                          {
-                            outputs: {
-                              address: {
-                                _in: $addresses
-                            }
-                          }}
-                        ]
-                      }) {
-                        inputs {
-                          address
-                        }
-                        outputs {
-                          address
-                        }
-                      }
-                    }
-                  `;
-    console.log(JSON.stringify({query, addresses}))
-    return axios.post(graphqlEndpoint,
-                     JSON.stringify({query, addresses}),
-                     contentTypeHeaders);
-};
-
-const askBlockNumByTxHash = async (hash : string): Promise<utils.UtilEither<number>> => {
-    const query = `
-            query BlockNumByTxHash($hashId: Hash32HexString!) {
-              transactions(
-                where: {
-                  hash: {
-                    _eq: $hashId
-                  }
-                }
-              ) {
-                hash
-                block {
-                  number
-                }
-              }
-            }`;
-    const ret = (await axios.post(graphqlEndpoint,
-                      JSON.stringify({ 'query': query, 'variables': {'hashId':hash} }),
-                      contentTypeHeaders));
-    if('data' in ret 
-       && 'data' in ret.data 
-       && 'transactions' in ret.data.data
-       && ret.data.data.transactions[0] 
-       && 'block' in ret.data.data.transactions[0]
-       && 'number' in  ret.data.data.transactions[0].block
-       && typeof ret.data.data.transaction[0].block.number === 'number')
-       return {kind:'ok', value:ret.data.data.transaction[0].block.number};
-    else 
-        return {kind:'error', errMsg: 'Did not understand graphql response'};
-
-
-} ;
-
-const askBlockNumByHash = async (hash : string) : Promise<utils.UtilEither<number>> => {
-    const query = `
-            query BlockNumByHash($id: Hash32HexString!) {
-              blocks(
-                where: {
-                  id: {
-                    _eq: $id
-                  }
-                }
-              ) {
-                id
-                number
-              }
-            }
-    `;
-    const ret = (await axios.post(graphqlEndpoint,
-                      JSON.stringify({ 'query': query, 'variables': {'hashId':hash} }),
-                      contentTypeHeaders));
-    if('data' in ret 
-       && 'data' in ret.data 
-       && 'transactions' in ret.data.data
-       && ret.data.data.transactions[0] 
-       && 'block' in ret.data.data.transactions[0]
-       && 'number' in  ret.data.data.transactions[0].block
-       && typeof ret.data.data.transaction[0].block.number === 'number')
-       return {kind:'ok', value:ret.data.data.transaction[0].block.number};
-    else 
-        return {kind:'error', errMsg: 'Did not understand graphql response'};
-
-};
-
-const askUtxoForAddresses = (addresses: string[]) => {
-    const query = `
-            query UtxoForAddresses($addresses: [String]) {
-              utxos(where: {
-                address: {
-                  _in: $addresses
-                }
-              }) {
-                address
-                txHash
-                index
-                value
-                utxo_tx {
-                      block { number }
-                }
-              }
-            }`;
-    return axios.post(graphqlEndpoint, 
-                      JSON.stringify({query, addresses}),
-                      contentTypeHeaders);
-};
-
-const askUtxoSumForAddresses = (addresses: string[]) => {
-    const query = `
-            query UtxoForAddresses($addresses: [String]) {
-              utxos_aggregate(where: {
-                address: {
-                  _in: $addresses
-                }
-              }) {
-                aggregate {
-                  sum {
-                    value
-                  }
-                }
-              }
-            }
-    `;
-    return axios.post(graphqlEndpoint,
-                      JSON.stringify({ 'query': query, 'variables': {'addresses':addresses} }),
-                      contentTypeHeaders);
-};
-
-interface TransactionFrag {
-    id: string;
-    block: BlockFrag;
-    includedAt: Date;
-    inputs: TransInputFrag;
-    outputs: TransInputFrag; // technically a TransactionOutput fragment
-}
-interface BlockFrag {
-    number: number;
-    hash: string;
-    epochNo: number;
-    slotNo: number;
-}
-interface TransInputFrag {
-    address: string;
-    value: string;
-}
-
-const askTransactionHistory = async ( 
-           limit: number
-         , addresses: string[]
-         , afterNum: utils.UtilEither<number>
-         , untilNum: utils.UtilEither<number>) : Promise<utils.UtilEither<TransactionFrag[]>> => {
-    const query = `
-            query TxsHistory($addresses: [String], $limit: Int, $afterBlockNum: Int, $untilBlockNum: Int) {
-              transactions(
-                where: {
-                  _and:
-                  [
-                    {
-                      block: {
-                        number: {
-                          _gte: $afterBlockNum,
-                          _lte: $untilBlockNum
-                        }
-                      }
-                    },
-                    {
-                      _or: [
-                    {
-                      inputs: {
-                        address: {
-                          _in: $addresses
-                        }
-                      }
-                    },
-                    {
-                      outputs: {
-                        address: {
-                          _in: $addresses
-                        }
-                      }
-                    }
-                  ]
-                    }
-                  ]
-                },
-                limit: $limit,
-                order_by: {
-                  includedAt: desc
-                }
-              ) {
-                id
-                block {
-                  number
-                  hash
-                  epochNo
-                  slotNo
-                }
-                includedAt
-                inputs {
-                  address
-                  value
-                }
-                outputs {
-                  address
-                  value
-                }
-              }
-            }
-    `;
-    const ret = await axios.post(graphqlEndpoint,
-                      JSON.stringify({ 'query': query,
-                                       'variables': {'addresses':addresses,
-                                                     'limit': limit,
-                                                     'afterBlockNum': afterNum.kind === 'ok'? afterNum.value : null,
-                                                     'untilBlockNum': untilNum.kind === 'ok'? untilNum.value : null} }),
-                      contentTypeHeaders);
-    if('data' in ret && 'data' in ret.data && 'transactions' in ret.data.data)
-        return {'kind':'ok', value:ret.data.data.transactions};
-    else
-        return {'kind':'error', errMsg:'TxsHistory, could not understand graphql response'};
-
-
-};
-
-const filterUsedAddresses = async (req: Request, res: Response) => {
-  if(!req.body || !req.body.addresses) {
-      console.log("error, no addresses.");
       return;
-  }
-  const verifiedAddresses = utils.validateAddressesReq(addressesRequestLimit
-                                                      , req.body.addresses);
-  switch(verifiedAddresses.kind){
-      case "ok": 
-          const result = (await askFilterUsedAddresses(verifiedAddresses.value)).data.data;
-          const usedAddresses = _.chain(result.transactions)
-                                 .flatMap(tx => [...tx.inputs, ...tx.outputs])
-                                 .map('address')
-                                 .intersection(verifiedAddresses.value)
-                                 .value();
-
-          res.send(usedAddresses);
-          return;
-      case "error":
-          console.log(verifiedAddresses.errMsg);
-          return;
-      default: return utils.assertNever(verifiedAddresses);
-  }
+    case "error":
+      console.log(result.errMsg);
+      return;
+    default: return utils.assertNever(result);
+  };
 };
-
 
 const utxoForAddresses = async (req: Request, res: Response) => {
   if(!req.body || !req.body.addresses) {
@@ -328,19 +59,60 @@ const utxoForAddresses = async (req: Request, res: Response) => {
                                                       , req.body.addresses);
   switch(verifiedAddresses.kind){
       case "ok": 
-          const result = (await askUtxoForAddresses(verifiedAddresses.value)).data.data;
-          const utxos = Array.isArray(result)
-                        ? result.map( utxo => 
-                                        ({
-                                          utxo_id: `${utxo.txId}:${utxo.index}`,
-                                          tx_hash: utxo.txHash,
-                                          tx_index: utxo.index,
-                                          receiver: utxo.address,
-                                          amount: utxo.value,
-                                          block_num: utxo.utxo_tx.block.number,
-                                        }))
-                        : []
-          res.send(utxos)
+          const result = await askUtxoForAddresses(verifiedAddresses.value);
+          switch(result.kind)
+          {
+            case "ok":
+              const utxos = result.value.map( utxo => 
+                                            ({
+                                              utxo_id: `${utxo.txHash}:${utxo.index}`,
+                                              tx_hash: utxo.txHash,
+                                              tx_index: utxo.index,
+                                              receiver: utxo.address,
+                                              amount: utxo.value,
+                                              block_num: utxo.transaction.block.number,
+                                            }));
+              res.send(utxos);
+              return;
+            case "error":
+              console.log(result.errMsg);
+              return;
+            default: return utils.assertNever(result);
+
+          }
+      case "error":
+          console.log(verifiedAddresses.errMsg);
+          return;
+      default: return utils.assertNever(verifiedAddresses);
+  }
+};
+
+
+const filterUsedAddresses = async (req: Request, res: Response) => {
+  if(!req.body || !req.body.addresses) {
+      console.log("error, no addresses.");
+      return;
+  }
+  const verifiedAddresses = utils.validateAddressesReq(addressesRequestLimit
+                                                      , req.body.addresses);
+  switch(verifiedAddresses.kind){
+      case "ok": 
+          const result = await askFilterUsedAddresses(verifiedAddresses.value);
+          switch(result.kind){
+            case "ok":
+              const usedAddresses = _.chain(result.value)
+                                     .flatMap(tx => [...tx.inputs, ...tx.outputs])
+                                     .map('address')
+                                     .intersection(verifiedAddresses.value)
+                                     .value();
+
+              res.send(usedAddresses);
+              return;
+            case "error":
+              console.log(result.errMsg);
+              return;
+            default: return utils.assertNever(result);
+          }
           return;
       case "error":
           console.log(verifiedAddresses.errMsg);
@@ -348,6 +120,8 @@ const utxoForAddresses = async (req: Request, res: Response) => {
       default: return utils.assertNever(verifiedAddresses);
   }
 };
+
+
 
 const utxoSumForAddresses = async (req:  Request, res:Response) => {
   if(!req.body || !req.body.addresses) {
@@ -358,8 +132,16 @@ const utxoSumForAddresses = async (req:  Request, res:Response) => {
                                                       , req.body.addresses);
   switch(verifiedAddresses.kind){
       case "ok": 
-          const result = (await askUtxoSumForAddresses(verifiedAddresses.value)).data.data;
-          res.send(result);
+          const result = await askUtxoSumForAddresses(verifiedAddresses.value);
+          switch(result.kind) {
+            case "ok":
+              res.send({ sum: result.value });
+              return;
+            case "error":
+              console.log(result.errMsg);
+              return;
+            default: return utils.assertNever(result);  
+          }
           return;
       case "error":
           console.log(verifiedAddresses.errMsg);
