+++ src/services/bestblock.ts
@@ -1,0 +1,43 @@
+import axios from 'axios';
+import { Request, Response } from "express";
+
+import { assertNever, contentTypeHeaders, graphqlEndpoint, UtilEither} from "../utils";
+
+interface CardanoFrag {
+  blockHeight: number;
+  currentEpoch: EpochFrag;
+  slotDuration: number;
+}
+
+interface EpochFrag {
+    blocks: BlockFrag[];
+    number: number;
+}
+
+interface BlockFrag {
+    hash: string;
+    number: number;
+}
+
+export const askBestBlock = async () : Promise<UtilEither<CardanoFrag>> => {
+    const query = `
+                {
+                  cardano {
+                    blockHeight,
+                    currentEpoch {
+                      number
+                      blocks(limit:1, order_by: { createdAt:desc}) {
+                        hash
+                        number
+                      }
+                    },
+                    slotDuration,
+                  },
+                }
+            `;
+    const ret = await axios.post(graphqlEndpoint, JSON.stringify({'query':query}), contentTypeHeaders);
+    if('data' in ret && 'data' in ret.data && 'cardano' in ret.data.data)
+        return { kind: 'ok', value: ret.data.data.cardano };
+    else return { kind: 'error', errMsg:'BestBlock, could not understand graphql response' };
+};
+
