+++ src/services/utxoSumForAddress.ts
@@ -1,0 +1,35 @@
+import axios from 'axios';
+import { Request, Response } from "express";
+
+import { assertNever, contentTypeHeaders, graphqlEndpoint, UtilEither} from "../utils";
+
+export const askUtxoSumForAddresses = async (addresses: string[]): Promise<UtilEither<string>> => {
+    const query = `
+            query UtxoForAddresses($addresses: [String]) {
+              utxos_aggregate(where: {
+                address: {
+                  _in: $addresses
+                }
+              }) {
+                aggregate {
+                  sum {
+                    value
+                  }
+                }
+              }
+            }
+    `;
+    const ret = await axios.post(graphqlEndpoint,
+                        JSON.stringify({ 'query': query, 'variables': {'addresses':addresses} }),
+                        contentTypeHeaders);
+    if(   'data' in ret && 'data' in ret.data 
+       && 'utxos_aggregate' in ret.data.data 
+       && 'aggregate' in ret.data.data.utxos_aggregate
+       && 'sum' in ret.data.data.utxos_aggregate.aggregate
+       && 'value' in ret.data.data.utxos_aggregate.aggregate.sum)
+      return { kind: 'ok', value: ret.data.data.utxos_aggregate.aggregate.sum.value };
+    else
+      return { kind: 'error', errMsg: 'utxoSumforAddresses, could not unstand graphql response.'};
+
+
+};
